extracted tests and pcode_op_simple module
